Allow any Ruby 2.7.x version
Background: Out hosting provider maintains installed Ruby version and installs patch versions shortly after they are released. This is good for as, as e.g. important security patches are installed for us.

However, it may leed to (and has led to) issues, as bundler checks for the installed ruby version when running `bundle exec` and fails if it is not available.

As generally patch versions are backwards compatible, it’s safe to use this slightly more pessimistic version constraint in this case.

Signed-off-by: Till Prochaska <mail@tillprochaska.de>