Show error view in case of any errors
@mattwr18 I wanted to make sure the error view is rendered even if no JWT is passed at all. In that case `ActionController::ParameterMissing` would be raised.

I could have simply added that exception to the `rescue` line, but decided to simply rescue any errors. I think it’s fair to display the helpful error message in case of any error, even if it might not be directly related to the token. After all, the user needs to know what to do next (i.e. contact support).

Signed-off-by: Till Prochaska <mail@tillprochaska.de>