Favor telegram_id over telegram_chat_id
- we don't receive the chat_id from the telegram widget, but looking
  closer the chat_id and the id are the same for individual users, like
our use case. We checked the data in our old staging server as well and
this help up. We have not written a migration to remove the
telegram_chat_id from the Contributor model because it's safer to leave
it in case we are off on this. The docs aren't super clear about this.