refactor: prefer node lts version
Current long term support version of node is v14. I'd prefer to point
all our nodejs version constraints to it.

For alpine linux v3.11 (what we're using for the docker base image), the
LTS version is still v12 but I'm sure it will update when the dockerfile
updates.
https://pkgs.alpinelinux.org/packages?name=nodejs&branch=v3.11

I checked it with:
```
$ docker-compose exec app node --version
v12.20.1
```

I'm using asdf-vm as a global version manager. It uses `.tool-versions`.
Unfortunately it does not have a symlink to LTS yet:
https://github.com/asdf-vm/asdf-nodejs/pull/176