Refactor: Move invite url creation to own controller
As the invite creation for new contributors is currently the only action requiring authentication in the onboarding controller and also the only action which is done by users and not used during the onboarding workflow, it doesn't really belong in the onboarding controller. Additionally, keeping it in the onboarding controller increases the danger of introducing a bug where no authentication is needed for creating invite urls.