Add interstitial component
This component serves as a generic component to display some kind of info in an interstial, i.e. in a small, centered box on the side, usually outside of the classic application layout.

I’m using it for the success/error screens as well as Threema/email forms in the onboarding process. However, it could be used in other contexts as well.