Simplify onboarding instructions/channel selection
We now simply display a list of buttons for each channel (Threema, Telegram, email) that links to a separate page displaying all the information as well as forms etc. for the respective channel.

This simplifies the amount of custom JS/markup required a lot, and it allows us to add new channels to the onboarding process more easily.