Add translation helper to ApplicationComponent
When used within a component, the `t` helper now automatically looks for a translation within the component scope, e.g. given a component `MyComponent`, calling `t 'my_translation'` will try to find `components.my_component.my_translation` first.

If such a scoped translation does not exist, it’ll return the value of a global translation, i.e. in the example above, it would try to find the `my_translation` key in the translation file.

As we’re adhering to the convention of keeping our component-specific translations scoped like this, this saves us from a lot of redundancy, and is backwards compatible in cases where we don’t follow the convention.

(The only edge case is if both, a scoped and a global translation exist for a key. However, I don’t think there are any real cases where the logic of this helper does not lead to the desired result, so I think this should be fine.)