feat: strip previous text from message
The biggest change here is that we use the decoded `html_part` of
incoming emails if it exists. The rationale behind this is that we can
properly parse incoming messages and don't get a text blob. Thus, we're
able to remove the part of the e-mail that contained or previous
message. I think it should be safe, as we treat and santize incoming messages
as HTML anyways.