Unify submit buttons in contributor form
Previously, we used a submit and reset button in the contributors form that’s been left over from one of the first UI iterations.

In order to unify form design with other forms in 100eyes, I removed the reset button, so the form now has a single, primary action button.