Destroy Tagify instance
Destroying the respective Tagify instance when the coresponding DOM element is removed makes sure we don’t keep unnecessary objects around.

In this particular case it’s necessary, as Turbo caches the DOM for pages previously visited. If we do not destroy the Tagify instance, the page would be cached *including*. an initialized tags input. If the cached page was then rerendered, trying to initialize Tagify again would cause an error.