Remove notifications before caching page
If a user visits a page they have previously visits, Turbo tries to render a cached version of the page while loading an up to date version from the server.

However, notification shouldn’t be displayed when a cached instance of the page is displayed.

Turbo caches pages directly before navigating away, so we can safely remove them before caching the page.