Add skip links
When navigation web pages using the keyboard, having to tab over the same (navigational) elements on every page quickly becomes annoying. Skip links are a common solution.

We should have added a skip link to skip the nav bar when it was added, but I think it’s also necessary to be able to skip the list of contributors on profile pages.