refactor: change API of telegram inbound adapter
This is a major refactoring. The reason for this change is that we want
the control flow in the telegram webhook controller. The telegram
adapter should be able to process a telegram message and execute
callbacks on certain events e.g. when a contributor is connected or not
found at all.