Reduce code duplication in `CopyButton` and `CreateInviteButton` components
Before, we had two different components: A generic `CopyButton` that would copy any text you’d pass into it on click, and a `CreateInviteButton`, that would fetch an invite link from a URL, then write it to the clipboard.

Except for the text to copy (static text passed in via props vs. fetched dynamically from a URL), both behaved exactly the same.

I decided to merge both components, so we now have a single `CopyButton` component that handles both use cases.

This example writes the text “Lorem ipsum” to the clipboard:

```ruby
<%= c 'copy_button', label: 'Copy me', copy: 'Lorem Ipsum' %>
```

This example fetches the URL `/example` and writes the contents of the `text` key of the JSON response to the clipboard.

```ruby
<%= c 'copy_button', label: 'Copy me', url: '/example', key: 'text' %>
```