Use slightly less spacing between icon and text
Before, the gap between the icon and the text and between multiple elements of the icon list where the same. In some cases, this can make it *slightly* more difficult to visually associate which icons and texts belong to each other. Making the gap between the icon and text of an individual elements slightly smaller than the gap between multiple elements helps.