WIP Add BaseField component
The current `Field` component does not only render form fields – it tries to determine field labels, help texts etc. automatically given an ActiveRecord object.

This is convenient in many cases, as most of the time we’re displaying forms related to some record. However, in cases where we want to display fields not directly related to a record (e.g. sign in and password reset forms), this requires workarounds.

The new `BaseField` component doesn’t do any of the magic – it simply accepts a label, help text, errors etc. as properties and renders the field. In cases where we simply want to render a form field, we can now use this component.

As a second step, I’m going to refactor the existing `Field` component to use the `BaseField` under the hood.