Use components in sing-in/pass reset forms
Due to the previous limitations of the `Field` component, they weren’t easy to use in some cases like sign-in forms etc. This patch refactors those forms to use the new `BaseField` component (and also slightly updates the form design).

Apart from that, this commit introduces a few UX improvements:

* Sets autofocus to the first input in sign in, password reset forms etc.
* Uses the appropriate `autofocus` hints to enable browsers/password managers to automatically fill in inputs
* Uses the `password_input` component to allow users to unmask the password input instead of entering the password twice.
* Uses native HTML form validations to provide feedback on e.g. password length, required inputs etc.