Add splash screen prompting users to send an initial Signal message
In order to reliably send messages to contributors via Signal, they need to contact us first. Otherwise, Signal seems to silently drop our messages. We assume this is a spam protection strategy.

This could be further improved: After a Signal an unknown sends us an initial message, we should reply to them with a link that allows them to continue onboarding, e.g. something along the lines of:

> Hey, welcome to [project name]! Please complete your registration here: https://project.100ey.es/onboarding/signal

Currently, the onboarding token is passed on throughout the onboarding flow via a URL parameter. In order to implement a welcome message like this, we would need to store it in the session. (That might be a good idea anyways and results in cleaner links.)