Add dedicated queue for polling signal messages with lower priority
Requests against the signal-cli REST API can take quite long (20 - 30 seconds)
from our experience. As we also poll Signal messages once per minute, this
could potentially lead to blocking the sending of messages to contributors,
which previously were enqueued with jobs on the same queue.
We now have a dedicated queue for Signal message polling with a lower priority
to avoid this.