refactor: remove foreman for easier debugging
Using `binding.pry` a docker environment gets a pain in the ass once
foreman enters the game.

@tillprochaska: As far as I can see it, there is no hot-reload feature
anymore. But I checked out your last commit and there it seems not to be
there as well. So, I guess, I didn't break anything with my refactoring.

The multiple docker-services seem to work fine. I remember that we had
an issue with shared volumes between the `assets` service and the `app`
service in the past. This does not seem to be an issue anymore (at least
on my machine).