Reset button loading state before caching page
Turbo caches the current page contents before leaving the page in order to immediately show it if the user e.g. navigates back.

In order to provide direct feedback when a user submits a form, we display a loading spinner and disable the submit button to prevent duplicate form submissions.

We need to reset this before caching the page contents.