Remove JWT requirement for Telegram link route
In order to support contributors in completing the Telegram onboarding flow, we want to be able to (re-)send the link that contains the instructions to contributors.

Currently, the route is protected by a JWT check, that I have now removed.

In order to not disclose wether a Telegram onboarding token is valid, the route now simply displays the token as it is encoded in the URL (and doesn’t validate it anymore).