Implement POC of two-step sign-in
Right now, we ask users to provide their email and password as well as the OTP in the same step. This is problematic from a UX point of view:

1) We ask users to provide a OTP even if they haven’t yet set up a 2FA.

2) Feedback can be confusing because useres don't know wether it was the email/password combination or the OTP they entered incorrectly.

This commits implements a "true" two-step sign-in process. Users are asked for their email/password, and if both are correct *and* 2FA is set up, prompts for the OTP.

I’m unsure about the implementation of OTP verification:

Before, we would only sign in a user after they provided their email/password and OTP. The downside of this approach was that we had to overwrite quite a lot of Clearance’s logic, because there was no way to hook into the auth process in a clean way.

The new implementation let’s Clearance handle sign in without checking for an OTP, i.e. a speaking technically, a user is already signed-in after providing email/password, even if they have 2FA set up.

In a global `before_action` we then check if the user has 2FA set up and has not yet provided their OTP. Once they provide a valid OTP, we set a session variable `otp_verified_for_user` to the user’s id so we don’t keep asking for an OTP.