Show hint to use native emoji picker
Emoji are often used in personal, casual text messages. The latest versions of both Windows and macOS support a native emoji picker, but not everyone knows they exist.

So instead of implementing our own emoji picker or embedding a huge library, we simply display a small hint below textareas to make our users aware of the native picker if they are on a supported platform.

Closes #953