Fix: Render issue after failed avatar update.
Previously, when a contributor avatar was updated with a not allowed file type,
active storage validator correctly signaled an invalid update. However, this only
stopped the invalid file from being persisted in the database, the `@contributor`
object was still updated (due to the way the Rails `update` implementation works).
As we directly render the show view using the `@contributor` object from the `#update` action,
this then failed because the `@contributor` object still contained an invalid avatar file blob.
Resetting `@contributor` from the database after a failed update resolves this issue.