Refactor settings form
The implementation of the settings form has been quite verbose and a little difficult to read because of the nested field and input components. I’ve introduced a new `SettingsField` component that implements this so we can now add new fields to the settings form in a single line, which makes it a lot easier to read/maintain imo.