Clear settings cache before performing jobs
Fix #1276. This is the simplest and most generic solution I could come up with.

Instead of registering a DelayedJob lifecyle callback, we could have used `ActiveJob`’s `before_perform` callback. However, this would require some code duplication in order to solve this problem for any async job (mail delivery and other jobs), as we’d have to implement the callback method in both `ApplicationJob` and a subclass of `ActionMailer::MailDeliveryJob`.

Additionally, this would clear the cache any time a job is performed – even if the job is performed syncronously. (This isn’t relevant right now, but it might be unexpected in case we ever use sync jobs in the future.)