Add admin dashboard for delayed jobs
Currently, we try to execute a job at most once, but we do keep failed jobs in the database. We currently do not make use of this, but I think it would be very helpful to be able to see which jobs failed (and why), and (in the future) even to manually retry a failed job.