Add handling of signal typing messages.
Previously, we handled supported messages by explicitly checking if
a message was of an unsupported type. As all messages that we currently
want to handle are of the 'dataMessage' type, I changed the check to
excluding / not handling anything that is not a 'dataMessage'.
Hopefully that will help us avoid issues with other message types
which we don't support.