Remove hints feature
We introduced a feature to append text blocks/canned messages (for some reason named “hints) to requests, but later found out that we didn’t have a good use case for it. We already removed it from the UI some time ago. I’m removing the backend implementation/specs now as well, because it’s unlikely we’ll need this feature again.

We have considered adding a new feature that allows editors to quickly reply using canned replys/text blocks, but that would likely be implemented differently.

Close #389