Fix highlighting in textareas
The previous implementation had a problem: The cursor wouldn’t be visible if inside a highlighted section, because the highlights where layered on top of the actual textarea.

This reverses the order: Now, the textarea is placed on top of the highlighted text. However, everything except for the textarea cursor indicator is invisible (i.e. it has transparent text color).