Fix sending emails with commas in their sender name.
To comply with RFC 5322 (https://www.rfc-editor.org/rfc/rfc5322#section-3.6.2) where commas in the from: field specify a list of multiple senders, we need to escape commas in the sender name. This is necessary if a project specifies its project name and therefore the sender name, e.g. as 'Coffee, and you?'.