Display raw message data in admin dashboard
This displays the raw message data in the admin dashboard in order to help debugging messages that contain unsupported content types etc.

We store the raw message data for all inbound messages as ActiveStorage attachments. administrate doesn’t support attachments out of the box. There are gems that add support, but in this case, all we need is to display a list of links to those attachments, which can be implemented without too much effort using a custom administrate field type.

Closes #462