Disable contributors edit form in admin dashboard
Editing contributors via the admin dashboard isn’t possible, but the edit form for contributors is accessible via an “Edit” button in the admin dashboard. This removes the respective `edit` route. As administrate automatically displays/hides action buttons (like the edit button) depending on wether a respective route is available, this will also hide the “Edit” button in the admin dashboard.