Improve Threema ID input UX
Technically, we do not care about the casing of the Threema ID anymore. However, the Threema app displays the ID in uppercase letters so most users will try to enter it in uppercase letters.

This commit automatically displays all text entered into Threema ID inputs in uppercase letters, so users are not confused about wether case is relevant. It also saves users (on mobile) a few keystrokes/touches as they don’t have to press the shift key.