Add contributor contact creation to Signal onboarding workflow.
Signal message senders are now being identified by UUIDs instead of by phone number. To still be able to obtain the phone number of a message sender (which we need to match incoming messages to contributors) we need to create a contact for each contributor. The UUID and phone numbers are then being resolved automatically by signal-cli / Signal.