Remove commented out expectations
- we don't get any network traffic in a real browser when we open a new
  window with target="_blank", which means we don't have the
page.response_headers. If we remove the target="_blank" we can add these
tests back, if they are important, but that would change the UX. Note
that forcing the user to open a new window is an a11y, UX, and security
concern. Though for the latter, this would not be relevant as we are not
opening a third party site.