ci: check for unstaged files
When you do the following
```
$ asdf install
$ npm --version
6.14.16
```

and run:

```
npm install
```

~then you get a changed `package-lock.json`.~

It is good practice to always check in lock files into version control.
On CI we run `npm install` (among other things). If we make sure that we
do not leave any updates files, this ensures we always commit our lock
files.

EDIT: You don't, my fault. But it does not hurt to merge this PR anyways.