fix: handle messages without text or caption
Motivation
----------
If our webhook controller responds with 500 then Telegram API will try to send the message over and over.

We have a particular message that we did not expect: It's a message without text (which can happen) and we receive it before the user has onboarded herself. E.g you could simulate this by sending a sticker to `MAZ_Potsdam_Bot`.

How to test
-----------
1. `bin/rspec ./spec/adapters/telegram_adapter/inbound_spec.rb`
2. See:
```
  #consume
    |message| block argument
      ...
      given the payload does not have a caption nor text
        is expected to be nil
        when UNKNOWN_CONTRIBUTOR is registered
          is expected to yield control
```