fix: broken invitation link
merge after #1728

Motivation
----------
So this is a follow up for #1728.

It will prevent rendering the whatsapp link unless activated.

We still have an inconsistency, because the channel links depend on
`Setting.channels` whereas the routing depends on
`Setting.whatsapp_configured?`. Ie. `Setting.channels` can be activated
for whatsapp although `Setting.whatsapp_configured?` is false.

How to test
-----------
1. `bin/rspec spec/system/onboarding/index.spec.rb`
2. See:
```
Onboarding
  visit /onboarding/
    if WhatsApp was explicitly activated or activated by configuration
      renders invitation link for WhatsApp
    but if WhatsApp is deactivated
      renders no invitation link for WhatsApp
```