Favor using contributor id over sender id
- Since the request feed is scoped to the contributor and the anchor tag
  is specific that it expects a contributor id it seems to more semantic
to use the contributor id over the sender id. Senders can be either
contributors or users at present. If in the future, we want to use the
request feed component to link to user messages as well, we will need to
refactor the component to display the sender name and update the anchor
tag to avoid assuming the sender is a contributor.

It does not seem that there is an additional db query to use the
sender.id, perhaps because the message has the `sender_id` and there is
an index on the messages; however, as this is the only attr of the
sender being used and we know that the sender == contributor, it seems
safer to just use the contributor.id.