Fix specs failing when run in docker
- We mock the value of Setting.signal_server_phone_number, but were
  filtering out the value of ENV.fetch('SIGNAL_SERVER_PHONE_NUMBER',
nil). Filtering out the env variables themselves is flaky because it
depends on local .env variables, which is different locally, either with
or without docker, and on different machines, like ci, or other teammates.

This PR proposes to filter out the value of
Setting.signal_server_phone_number, which would be the same as the
request that is being received by vcr.