Only mount the app directory for assets service
We want to mount the directory to benefit from esbuild/postcss watch
functionality so that changes to our local dev code can be re-built and
available to refresh the page and be applied when using docker. This
allows us to not need to manually run the build commands every time a
change is made or stop/start the containers again.

We do not want the default volume mount because then the node_modules
reflect whatever we have locally and do not benefit from the
node_modules created during npm install in the build phase.