refactor: ansible backup for managed cluster
Motivation
----------
Almost all our deployments run on a managed cluster, as a matter of
fact, we want to use managed clusters only in the future.

The previous ansible configuration used the obsolete `db` docker
container only to connect to the managed postgres cluster, which is very
complicated.

This will remove the unnececcessary `docker exec ...` and also DRYs some
of the following configuration.

How to test
-----------
1. Follow the README on any production instance.
2. You should have a local backup in `ansible/backups/`