Add many-to-many relationship between users and organizations (#1971)
### What changed in this PR and why

This introduces the many-to-many relationship that is a requirement for
some of our clients to belong to more than one organization.

Instead of a simply HABTM, it uses `has_many, through:`. This gives us
more flexibility to save information on the relationship itself. For
example, how do we handle active/inactive state. One idea would be to
add a `deactivated_at` in the `users_organizations` table.


### TODOs

- [ ] Add automated tests that test for a user that belongs to multiple
organizations
- [x] Decide on the deactivated/reactivated logic
- [x] When an admin is creating a new user, try to find the user by
their email first and do something - ie either simply add the user to
the organization, or throw a validation error and direct the admin to
edit the existing user. EDIT: The email is already validated for
uniqueness by the clearance gem. An error is displayed if someone tries
to create a user with the same email.