Favor simply links for setting the organization
A select is favorably if the list is expected to be long, but if there
are only two or three items, the links give a better overview quickly.
Also, the page is not busy so compacting the list into a select does not
have the benefit it would normally. Finally, this removes the need for a
separate route/action to set the organization since we already do that
on every visit to any protected page on the site.