Remove global env variables from organizations (#2016)
Closes #1947 

@soey there is still the `three_sixty_dialog_partner_token` attribute
that lives at the organization level, which it probably shouldn't. This
is also a value that is global and is not unique to the organization.
The token is only good for 24 hours, which makes it difficult to get the
value from env variables. Previously, with `rails-settings-cached`, we
would save the value to the db in the `settings` table and then check to
see if there was a value and that it had been updated within 24 hours
before asking for a new token, or using the valid one.

Currently, this is used for creating the api key for a newly onboarded
client. It then is used to create the templates needed to send out
messages for that client.

I think a better way to handle this would be to assume the token is not
valid, if it even exists, since we are unlikely to be onboarding
multiple clients in the same 24-hour period and always ask for a new
token when we run the create api key job. The create template job is
currently called from a rake task that collects the templates from our
`de.yml` before scheduling the job. We would need to rethink this flow
if we wanted to be able to use the same token to create the templates.