Add spec for most common use case
we set the schedule_send_for to Time.current in the request form. Unless
the user removes the schedule_send_for manually or schedules it for the
future, the request schedule_send_for will be set to some time in the
near past. Checking in production, on one instance, there were actually
no cases where the user manually removed the schedule_send_for so it was
always present.