Remove data migrations workflow/gem (#2042)
These data migrations are meant to be run on production instances where
we need to manipulate the data once and not to be run when setting up
new instances. Also, if we skip a deploy on specific instances for some
reason, which is not ideal, but does happen, there are too many
considerations to be ideal.

We have a PR to make them more idempotent, but even so the first ones
were written when all instances were single tenant and we could be sure
there was only one organization, but after introducing multi-tenancy,
there is another layer that must be considered.