Fix create templates job failing on first run
Fixes #2044

The first time the job is run the
three_sixty_dialog_client_waba_account_id is nil so it calls the
fetch_client_info method. This method successfully updates the
organization, but as it's the last line of the method it saves the
return of the update `true` to the local variable that is used in the
url to create the templates. This results in an error. Subsequent times
it's ran it does not error because it uses the value of the
organization's attribute, which is the correct value.