Fix unscoped charts data (#2086)
We originally added the unscoped due to an error using the groupdate gem
and default scopes with postgresql. See
https://github.com/ankane/groupdate/issues/246

The solution was to remove all scopes because at the time the only scope
was the default order; however, this introduced a bug when we moved to
multi-tenancy and wanted to add the scope for the organization. The
cleaner sollution was always to remove only the order scope as is done
in this PR. See
https://github.com/ankane/groupdate?tab=readme-ov-file#default-scopes

The PR also adds tests that expose the bug and reorganize the tests to
meke them slightly more readable.

It adds a new scope that returns direct messages.

Fixes #2085