Set up log rotation for new instances (#2084)
There is also a play that has been tested for setting up log rotation
for existing instances. I have tested this for the roberta instance. I
have not tested the flow from the initial setup, which uses the same
task since I have not set up a new instance specifically for testing it.

Closes #2083