Refactor 360dialog inbound to reduce complexity
instead of the initialize_message method handling mutliple jobs, it now
handles one. There is a new method to check for ephemeral data, or
message text we receive that we do not want to create a message record
in the db for. There are several different texts the contributor can
send us that fits in this category. When we determine it is ephemeral,
we want to trigger the correct callback and return without building a
new message record even.