[fix] quick reply button not responding with the request (#2089)
Fixes #2081
Fixes #2093 
Fixes #2094

## What changed and Why

We now always trigger the request to receive request callback when we
receive the keyword. This fixes a bug where we would only send the first
request the contributor asked for. This happened cause we mark a
timestamp when we send out a template message, when a contributor
responds, we set it back to nil and mark a timestamp that they
responded.

This also involves several big refactorings. These include:

### Refactoring `WhatsAppAdapter::ThreeSixtyDialogInbound`
* It now handles its own "callbacks" as it seems more the responsibility
of the adapter to handle these errors/jobs to send out messages based on
some trigger than the ProcessWebhookJob whose job it should be to simply
accept the webhook and pass it on.
* `initialize_message` method to remove the multiple actions it was
triggering. This had several downsides: there were several checks to see
if the inbound message was ephemeral data and what type before
continuing. Now, we check the text/payload to see if the inbound message
is ephemeral, ie should not be persisted to the db before even
initializing the message instance.
* We now assign the request to the inbound message in the adapter
itself, which also handles persisting it.
* With this, we added the feature to support quote replying a previous
message. If we receive a context object with a message's external id,
which corresponds to a message we have saved, we know which request they
are responding to and don't need to assume it's the latest message they
received.

### Fix for the request to receive previous message not working

* We now have a new model `Message::WhatsAppTemplate`, which belongs to
a message record, whose job it is to save the external id of the
template message that was sent out. This allows us to find the message
by its whats_app_template that a contributor is requesting.

Before this PR, we never saved this external id when a message template
was sent out, so the feature never actually worked.

Originally, when I first started working on this fix, I just saved the
external_id to the message record itself, but this didn't sit right
since it was only a temporary external_id, which would get updated
when/if the contributor actually asked for the message. It has the
downside of not being able to properly keep track of delivery receipts.
Consider an example where we mark the message record as sent, received,
and read, but the contributor never actually asked for the message, we
just incorrectly marked it because we had the incorrect external id
assigned to it.

### Refactor of the `WhatsAppAdapter::ThreeSixtyDialogOutbound::Text`
job

* We move all the payloads to the place they are used. We also pass in a
type so that we can create the template message record after we know
that it was successfully sent out.

### Specs
* Missing specs were also added.